change data type

+++ go/common/ServicePoints.go
@@ -3,13 +3,13 @@ package common
 // Add a bool for transaction
 type IServicePoints interface {
 	// Register A Service Point, handler + service area
-	RegisterServicePoint(IServicePointHandler, int32) error
+	RegisterServicePoint(IServicePointHandler, uint16) error
 	// Handle a message and forward to the handler
 	Handle(IElements, Action, IVirtualNetworkInterface, IMessage, bool) IElements
 	// Handle a notification message, massage it to a change set and forward to the handler
 	Notify(IElements, IVirtualNetworkInterface, IMessage, bool) IElements
 	// Return the service point handler for the service name and area
-	ServicePointHandler(string, int32) (IServicePointHandler, bool)
+	ServicePointHandler(string, uint16) (IServicePointHandler, bool)
 }
 
 type IServicePointHandler interface {
